pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contextledger"
version = "0.1.0"
description = "Universal context layer and skill versioning platform for AI interfaces"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "mcp>=1.0",
    "sentence-transformers>=3.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]
openai = [
    "openai>=1.0",
]
postgres = [
    "psycopg2-binary>=2.9",
]
git = [
    "gitpython>=3.1",
]

[project.scripts]
ctx = "contextledger.cli.main:cli"
ctx-mcp = "contextledger.mcp.mcp_server:mcp.run"

[tool.pytest.ini_options]
testpaths = ["tests"]
